Test type hints and test get specific URL details
