+++ .pre-commit-config.yaml
@@ -5,17 +5,17 @@ repos:
     - id: black
       language_version: python3.11
 
-  - repo: https://github.com/pre-commit/mirrors-prettier
-    rev: "v3.1.0"
-    hooks:
-    - id: prettier
-      types: [javascript]
-
   - repo: https://github.com/pycqa/flake8
     rev: "7.0.0"
     hooks:
     - id: flake8
       args: [--extend-ignore, E501]
+      
+  - repo: https://github.com/pre-commit/mirrors-prettier
+    rev: "v3.1.0"
+    hooks:
+    - id: prettier
+      types: [javascript]
 
   - repo: https://github.com/pre-commit/mirrors-eslint
     rev: "v9.1.1"
