+++ src/utils/strings/model_strs.py
@@ -19,6 +19,7 @@ URL_ID = "urlID"
 URL_TITLE = "urlTitle"
 ADDED_BY = "addedBy"
 USERNAME = "username"
+IS_CREATOR = "isCreator"
 UTUB_DESCRIPTION = DESCRIPTION
 
 
@@ -42,3 +43,4 @@ class MODELS:
     URL = URL
     TAG_STRING = TAG_STRING
     UTUB_DESCRIPTION = UTUB_DESCRIPTION
+    IS_CREATOR = IS_CREATOR
